% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{biomaRt}
\alias{biomaRt}
\title{biomaRt}
\format{
A data frame with 42198 rows and 11 variables:
\describe{
  \item{\code{ensembl_gene_id}}{ENSEMBL gene id}
  \item{\code{chromosome_name}}{Chromosome name [1-22,X,Y]}
  \item{\code{start_position}}{start}
  \item{\code{end_position}}{end}
  \item{\code{description}}{Description}
  \item{\code{hgnc_symbol}}{HGNC symbol}
  \item{\code{ensembl_transcript_id}}{ENSEMBL transcript id}
  \item{\code{transcription_start_site}}{TSS}
  \item{\code{transcript_start}}{Transcript start}
  \item{\code{transcript_end}}{Transcript end}
  \item{\code{uniprotswissprot}}{UnitProt id}
}
}
\usage{
biomaRt
}
\description{
Curation of biomaRt
}
\details{
extraction using R.
}
\keyword{datasets}
\keyword{internal}
