% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blank_plot}
\alias{blank_plot}
\title{FUNCTION_TITLE}
\usage{
blank_plot(message)
}
\arguments{
\item{message}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_label}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}}
}
\keyword{internal}
