% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{caprion}
\alias{caprion}
\title{Caprion panel}
\format{
A data frame with 987 rows and 7 variables:
\describe{
  \item{\code{Protein}}{Protein name as in UniProt}
  \item{\code{Accession}}{UniProt id}
  \item{\code{Gene}}{HGNC symbol}
  \item{\code{Protein.Description}}{Detailed information on protein}
  \item{\code{GO.Cellular.Component}}{GO Ceullular component}
  \item{\code{GO.Function}}{GO function}
  \item{\code{GO.Process}}{GO process}
}
}
\usage{
caprion
}
\description{
Information based on Caprion pilot studies
}
\details{
See the Caprion repository involving its use.
}
\keyword{datasets}
