% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TwoSampleMR.R
\name{clump_data.args}
\alias{clump_data.args}
\title{FUNCTION_TITLE}
\usage{
clump_data.args(
  dat = NULL,
  clump_kb = 10000,
  clump_r2 = 0.001,
  clump_p1 = 1,
  clump_p2 = 1,
  pop = "EUR"
)
}
\arguments{
\item{dat}{PARAM_DESCRIPTION, Default: NULL}

\item{clump_kb}{PARAM_DESCRIPTION, Default: 10000}

\item{clump_r2}{PARAM_DESCRIPTION, Default: 0.001}

\item{clump_p1}{PARAM_DESCRIPTION, Default: 1}

\item{clump_p2}{PARAM_DESCRIPTION, Default: 1}

\item{pop}{PARAM_DESCRIPTION, Default: 'EUR'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\keyword{internal}
