% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_outcome_data.args}
\alias{extract_outcome_data.args}
\title{FUNCTION_TITLE}
\usage{
extract_outcome_data.args(
  snps = NULL,
  outcomes = NULL,
  proxies = TRUE,
  rsq = 0.8,
  align_alleles = 1,
  palindromes = 1,
  maf_threshold = 0.3,
  access_token = ieugwasr::check_access_token(),
  splitsize = 10000,
  proxy_splitsize = 500
)
}
\arguments{
\item{snps}{PARAM_DESCRIPTION, Default: NULL}

\item{outcomes}{PARAM_DESCRIPTION, Default: NULL}

\item{proxies}{PARAM_DESCRIPTION, Default: TRUE}

\item{rsq}{PARAM_DESCRIPTION, Default: 0.8}

\item{align_alleles}{PARAM_DESCRIPTION, Default: 1}

\item{palindromes}{PARAM_DESCRIPTION, Default: 1}

\item{maf_threshold}{PARAM_DESCRIPTION, Default: 0.3}

\item{access_token}{PARAM_DESCRIPTION, Default: ieugwasr::check_access_token()}

\item{splitsize}{PARAM_DESCRIPTION, Default: 10000}

\item{proxy_splitsize}{PARAM_DESCRIPTION, Default: 500}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[ieugwasr]{check_access_token}}
}
\keyword{internal}
