% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_data.args}
\alias{format_data.args}
\title{FUNCTION_TITLE}
\usage{
format_data.args(
  dat = NULL,
  type = "exposure",
  snps = NULL,
  header = TRUE,
  phenotype_col = "Phenotype",
  snp_col = "SNP",
  beta_col = "beta",
  se_col = "se",
  eaf_col = "eaf",
  effect_allele_col = "effect_allele",
  other_allele_col = "other_allele",
  pval_col = "pval",
  units_col = "units",
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  samplesize_col = "samplesize",
  gene_col = "gene",
  id_col = "id",
  min_pval = 1e-200,
  z_col = "z",
  info_col = "info",
  chr_col = "chr",
  pos_col = "pos",
  log_pval = FALSE
)
}
\arguments{
\item{dat}{PARAM_DESCRIPTION, Default: NULL}

\item{type}{PARAM_DESCRIPTION, Default: 'exposure'}

\item{snps}{PARAM_DESCRIPTION, Default: NULL}

\item{header}{PARAM_DESCRIPTION, Default: TRUE}

\item{phenotype_col}{PARAM_DESCRIPTION, Default: 'Phenotype'}

\item{snp_col}{PARAM_DESCRIPTION, Default: 'SNP'}

\item{beta_col}{PARAM_DESCRIPTION, Default: 'beta'}

\item{se_col}{PARAM_DESCRIPTION, Default: 'se'}

\item{eaf_col}{PARAM_DESCRIPTION, Default: 'eaf'}

\item{effect_allele_col}{PARAM_DESCRIPTION, Default: 'effect_allele'}

\item{other_allele_col}{PARAM_DESCRIPTION, Default: 'other_allele'}

\item{pval_col}{PARAM_DESCRIPTION, Default: 'pval'}

\item{units_col}{PARAM_DESCRIPTION, Default: 'units'}

\item{ncase_col}{PARAM_DESCRIPTION, Default: 'ncase'}

\item{ncontrol_col}{PARAM_DESCRIPTION, Default: 'ncontrol'}

\item{samplesize_col}{PARAM_DESCRIPTION, Default: 'samplesize'}

\item{gene_col}{PARAM_DESCRIPTION, Default: 'gene'}

\item{id_col}{PARAM_DESCRIPTION, Default: 'id'}

\item{min_pval}{PARAM_DESCRIPTION, Default: 1e-200}

\item{z_col}{PARAM_DESCRIPTION, Default: 'z'}

\item{info_col}{PARAM_DESCRIPTION, Default: 'info'}

\item{chr_col}{PARAM_DESCRIPTION, Default: 'chr'}

\item{pos_col}{PARAM_DESCRIPTION, Default: 'pos'}

\item{log_pval}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\keyword{internal}
