% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TwoSampleMR.R
\name{harmonise_data.args}
\alias{harmonise_data.args}
\title{FUNCTION_TITLE}
\usage{
harmonise_data.args(exposure_dat = NULL, outcome_dat = NULL, action = 2)
}
\arguments{
\item{exposure_dat}{PARAM_DESCRIPTION, Default: NULL}

\item{outcome_dat}{PARAM_DESCRIPTION, Default: NULL}

\item{action}{PARAM_DESCRIPTION, Default: 2}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\keyword{internal}
