% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{hg19}
\alias{hg19}
\title{hg19 information}
\format{
A data frame with 62559 rows and 8 variables:
\describe{
  \item{\code{chr}}{Chromosome [chr1-22,X,Y,...]}
  \item{\code{start}}{start}
  \item{\code{end}}{end}
  \item{\code{width}}{width}
  \item{\code{strand}}{strand}
  \item{\code{ENSEMBL}}{ENSMEBL gene}
  \item{\code{SYMBOL}}{HGNC symbol}
  \item{\code{UNIPROT}}{UniProt id}
}
}
\usage{
hg19
}
\description{
protein information
}
\details{
Curation from R
}
\keyword{datasets}
