% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{hg19Tables}
\alias{hg19Tables}
\title{hg19 Table}
\format{
A data frame with 19872 rows and 12 variables:
\describe{
  \item{\code{X.chrom}}{Chromosome}
  \item{\code{chromStart}}{chromStart}
  \item{\code{chromEnd}}{chromEnd}
  \item{\code{strand}}{Strand}
  \item{\code{acc}}{UniProt id}
  \item{\code{uniprotName}}{Protein}
  \item{\code{accList}}{List of UniProt ids}
  \item{\code{isoIds}}{isoIds}
  \item{\code{geneName}}{geneName}
  \item{\code{geneSynonyms}}{geneSynonyms}
  \item{\code{hgncSym}}{HGNC symbol}
  \item{\code{ensGene}}{ENSEMBL gene}
}
}
\usage{
hg19Tables
}
\description{
Gene information
}
\details{
Curation from UCSC.
}
\keyword{datasets}
\keyword{internal}
