% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{inf1}
\alias{inf1}
\title{Olink/INF1 panel}
\format{
A data frame with 92 rows and 9 variables:
\describe{
  \item{\code{uniprot}}{UniProt id}
  \item{\code{prot}}{Protein}
  \item{\code{target}}{Protein target name}
  \item{\code{target.short}}{Protein target short name}
  \item{\code{gene}}{HGNC symbol}
  \item{\code{chr}}{chromosome[1-13,16-17,19-22]}
  \item{\code{start}}{start}
  \item{\code{end}}{end}
  \item{\code{ensembl_gene_id}}{ENSEMBL gene}
}
}
\usage{
inf1
}
\description{
The panel is based on SCALLOP-INF
}
\details{
Assembled for SCALLOP-INF
}
\keyword{datasets}
