% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mr_forest_plot2}
\alias{mr_forest_plot2}
\title{MR forest plot}
\usage{
mr_forest_plot2(singlesnp_results, exponentiate = FALSE, alpha = 0.05)
}
\arguments{
\item{singlesnp_results}{Results based on single variants.}

\item{exponentiate}{Logic variable to indicate exponentiation, Default: FALSE.}

\item{alpha}{Confidence level, Default: 0.05.}
}
\value{
A graphic object.
}
\description{
Adaptation from TwoSampleMR::mr_forest_plot()
}
\seealso{
\code{\link[plyr]{dlply}},\code{\link[plyr]{mutate}}
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{geom_errorbarh}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{scale_manual}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{labs}}
}
\keyword{internal}
