% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mr_funnel_plot2}
\alias{mr_funnel_plot2}
\title{MR funnel plot}
\usage{
mr_funnel_plot2(singlesnp_results)
}
\arguments{
\item{singlesnp_results}{Results based on single variants.}
}
\value{
A graphic object.
}
\description{
Adaptation from TwoSampleMR::mr_funnel_plot()
}
\seealso{
\code{\link[plyr]{dlply}},\code{\link[plyr]{mutate}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{scale_manual}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{theme}}
 \code{\link[cowplot]{theme_cowplot}}
}
\keyword{internal}
