% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mr_leaveoneout_plot2}
\alias{mr_leaveoneout_plot2}
\title{MR leave-one-out analysis}
\usage{
mr_leaveoneout_plot2(leaveoneout_results, alpha = 0.05)
}
\arguments{
\item{leaveoneout_results}{Results from leave-one-out analysis.}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.05.}
}
\value{
A graphic object.
}
\description{
Adapatation from TwoSampleMR::mr_leaveoneout_plot()
}
\seealso{
\code{\link[plyr]{dlply}},\code{\link[plyr]{mutate}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{geom_errorbarh}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{scale_manual}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{margin}},\code{\link[ggplot2]{labs}}
}
\keyword{internal}
