% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mr_scatter_plot2}
\alias{mr_scatter_plot2}
\title{MR scatter plot}
\usage{
mr_scatter_plot2(mr_results, dat, alpha = 0.05)
}
\arguments{
\item{mr_results}{MR results.}

\item{dat}{data.}

\item{alpha}{Confidentce level, Default: 0.05.}
}
\value{
A graphic object.
}
\description{
Adaptation of TwoSampleMR::mr_scatter_plot()
}
\seealso{
\code{\link[plyr]{dlply}},\code{\link[plyr]{mutate}}
 \code{\link[TwoSampleMR]{mr_egger_regression}},\code{\link[TwoSampleMR]{mr_egger_regression_bootstrap}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_crossbar}},\code{\link[ggplot2]{geom_errorbarh}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{ggtheme}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{scale_manual}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{c("guide_bins", "guide_colourbar", "guide_coloursteps", "guide_legend", "guides", "guides")}},\code{\link[ggplot2]{guide_legend}}
 \code{\link[cowplot]{theme_cowplot}}
}
\keyword{internal}
