% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{Olink_NGS}
\alias{Olink_NGS}
\title{Olink/NGS panel}
\format{
A data frame with 1472 rows and 3 variables:
\describe{
  \item{\code{UniProt}}{UniProt id}
  \item{\code{Assay}}{Experimental assay}
  \item{\code{Panel}}{Olink panel}
}
}
\usage{
Olink_NGS
}
\description{
Information based on pilot studies
}
\details{
Curated from R.
}
\keyword{datasets}
