% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{Olink_qPCR}
\alias{Olink_qPCR}
\title{Olink/qPCR panels}
\format{
A data frame with 1112 rows and 7 variables:
\describe{
  \item{\code{UniProt}}{character COLUMN_DESCRIPTION}
  \item{\code{Panel}}{character COLUMN_DESCRIPTION}
  \item{\code{Target}}{character COLUMN_DESCRIPTION}
  \item{\code{gene}}{character COLUMN_DESCRIPTION}
  \item{\code{chr}}{character COLUMN_DESCRIPTION}
  \item{\code{start}}{integer COLUMN_DESCRIPTION}
  \item{\code{end}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
Olink_qPCR
}
\description{
Information on all qPCR panels
}
\details{
Curated from Excel.
}
\keyword{datasets}
