% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptideAssociationPlot.R
\name{peptideMapping}
\alias{peptideMapping}
\title{peptide-to-protein mapping}
\usage{
peptideMapping(protein, batch = "ZWK", mm = 5)
}
\arguments{
\item{protein}{protein name.}

\item{batch}{batch No. in the Caprion experiment.}

\item{mm}{maximum number of mismatches.}
}
\value{
a list containing mapping information.
}
\description{
peptide-to-protein mapping
}
\examples{
\dontrun{
  batch <- "ZWK"
  load(paste0("~/Caprion/pilot/",batch,".rda"))
  PROC <- pQTLtools::peptideMapping("PROC",mm=0)
}
}
