% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pQTLtools-package.R
\docType{package}
\name{pQTLtools-package}
\alias{pQTLtools}
\alias{pQTLtools-package}
\title{A protein Quantitative Trait Locus toolkit}
\description{
The seeds collection of data and utilties for (pQTL) analysis. At
this early stage, this repository collects information on a number of
protein panels, linking function for cis/trans classification, 2D
manhattan plots, 3D-plotly plots, forest plots among others availale
from R/gap; query results on genes, regions, and SNPs via
PhenoScanner, adding functionality to check for replication across
platforms and aspects of protein-related analysis such as
pQTL-Mendelian Randomization via TwoSampleMR, linkage through UniProt
IDs to other resources.
}
\details{
\tabular{ll}{
Package: \tab pQTLtools\cr
Title: \tab pQTL Tools\cr
Author: \tab Jing Hua Zhao\cr
Maintainer: \tab Jing hua Zhao \href{mailto:jinghuazhao@hotmail.com}{jinghuazhao@hotmail.com}\cr
License: \tab MIT + file LICENSE\cr
URL: \tab https://github.com/jinghuazhao/pQTLtools\cr
BugReports: \tab https://github.com/jinghuazhao/pQTLtools/issues\cr
}
}
\examples{
\dontrun{
# datasets
head(biomaRt)

# Olink-SomaLogic panel overlap
p <- list(setdiff(inf1$uniprot,"P23560"),
          setdiff(SomaLogic160410$UniProt[!is.na(SomaLogic160410$UniProt)],"P23560"))
cnames <- c("INF1","SomaLogic")
VennDiagram::venn.diagram(x = p, category.names=cnames,
                          filename='os.png', imagetype="png", output=TRUE)
m <- merge(inf1,SomaLogic160410,by.x="uniprot",by.y="UniProt")
u <- setdiff(with(m,unique(uniprot)),"P23560")
options(width=220)
o <- subset(inf1,uniprot \%in\% u)
dim(o)
o
vars <- c("UniProt","chr","start","end","extGene","Target","TargetFullName")
s <- subset(SomaLogic160410[vars], UniProt \%in\% u)
dim(s)
us <- s[!duplicated(s),]
dim(us)
us

# SCALLOP/INF1
INF <- Sys.getenv("INF")
INF1_merge <- merge(inf1,
                    read.delim(file.path(INF,"work","INF1.merge-rsid"),as.is=TRUE),
                    by="prot")
INF1_uniprot <- unique(with(INF1_merge,uniprot))

# INTERVAL SomaLogic at box
HOME <- Sys.getenv("HOME")
box <- read.delim(file.path(HOME,"SomaLogic","doc","INTERVAL-box.tsv"),as.is=TRUE)
box_INF1 <- subset(box,UniProt \%in\% INF1_uniprot)
box_uniprot <- setdiff(unique(with(box_INF1,UniProt)),"P23560")
setdiff(INF1_uniprot,box_uniprot)

# Phenoscanner database
ps <- merge(subset(read.delim(file.path(INF,"work","pQTL_2018.txt.gz"),as.is=TRUE),
            pmid==29875488),
            box,by.x="trait",by.y="TargetFullName")
z <- subset(ps,UniProtgwas \%in\% INF1_uniprot & p<=1.5e-11)

# ST4 on Nature
st4regions <- subset(st4, UniProt \%in\% INF1_uniprot)
unique_uniprot_list <- setdiff(intersect(st4$UniProt,inf1$uniprot),"P23560")
subset(INF1_merge,uniprot \%in\% unique_uniprot_list)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jinghuazhao/pQTLtools}
  \item Report bugs at \url{https://github.com/jinghuazhao/pQTLtools/issues}
}

}
\author{
Jing Hua Zhao in collaboration with other colleagues
}
\keyword{internal}
