% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_coloc}
\alias{run_coloc}
\title{Colocalisation analysis}
\usage{
run_coloc(eqtl_sumstats, gwas_sumstats)
}
\arguments{
\item{eqtl_sumstats}{eQTL summary data.}

\item{gwas_sumstats}{GWAS summary data.}
}
\value{
Summary from \code{coloc.abf}.
}
\description{
The function takes eQTL and GWAS summary statistics for a colocalisation analysis..
}
