% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{SomaLogic160410}
\alias{SomaLogic160410}
\title{Somascan panel}
\format{
A data frame with 5178 rows and 10 variables:
\describe{
  \item{\code{SOMAMER_ID}}{Somamer id}
  \item{\code{UniProt}}{UniProt id}
  \item{\code{Target}}{Protein target}
  \item{\code{TargetFullName}}{Protein target full name}
  \item{\code{chr}}{chromosome[1-22,X,Y]}
  \item{\code{start}}{start}
  \item{\code{end}}{end}
  \item{\code{entGene}}{entrez gene}
  \item{\code{ensGene}}{ENSEMBL gene}
  \item{\code{extGene}}{external gene}
}
}
\usage{
SomaLogic160410
}
\description{
This is based on panel used in Sun et al. (2018) Nature.
}
\details{
from the INTERVAL study.
}
\keyword{datasets}
