% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{st18}
\alias{st18}
\title{Supplementary table 18}
\format{
A data frame with 3622 rows and 3 variables:
\describe{
  \item{\code{Number}}{A serial number}
  \item{\code{Analyte}}{Name}
  \item{\code{UniProt.ID(s)}}{UniProt id}
}
}
\usage{
st18
}
\description{
Supplementary information for Sun et al. (2018) Nature
}
\details{
As above.
}
\keyword{datasets}
