% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{st4}
\alias{st4}
\title{Supplementary table 4}
\format{
A data frame with 1980 rows and 31 variables:
\describe{
  \item{\code{Locus.ID}}{character COLUMN_DESCRIPTION}
  \item{\code{SOMAmer.ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Target}}{character COLUMN_DESCRIPTION}
  \item{\code{Target.fullname}}{character COLUMN_DESCRIPTION}
  \item{\code{UniProt}}{character COLUMN_DESCRIPTION}
  \item{\code{Sentinel.variant*}}{character COLUMN_DESCRIPTION}
  \item{\code{Chr}}{double COLUMN_DESCRIPTION}
  \item{\code{Pos}}{double COLUMN_DESCRIPTION}
  \item{\code{Region.start}}{double COLUMN_DESCRIPTION}
  \item{\code{Region.end}}{double COLUMN_DESCRIPTION}
  \item{\code{Effect.Allele.(EA)}}{character COLUMN_DESCRIPTION}
  \item{\code{Other.Allele.(OA)}}{character COLUMN_DESCRIPTION}
  \item{\code{EAF}}{double COLUMN_DESCRIPTION}
  \item{\code{INFO}}{double COLUMN_DESCRIPTION}
  \item{\code{cis/.trans}}{character COLUMN_DESCRIPTION}
  \item{\code{Mapped.gene}}{character COLUMN_DESCRIPTION}
  \item{\code{No..conditionally.significant.variants}}{double COLUMN_DESCRIPTION}
  \item{\code{Previously.reported}}{character COLUMN_DESCRIPTION}
  \item{\code{Replicates?}}{character COLUMN_DESCRIPTION}
  \item{\code{beta}}{double COLUMN_DESCRIPTION}
  \item{\code{SE}}{double COLUMN_DESCRIPTION}
  \item{\code{p}}{character COLUMN_DESCRIPTION}
  \item{\code{beta}}{double COLUMN_DESCRIPTION}
  \item{\code{SE}}{double COLUMN_DESCRIPTION}
  \item{\code{p}}{double COLUMN_DESCRIPTION}
  \item{\code{beta}}{double COLUMN_DESCRIPTION}
  \item{\code{SE}}{double COLUMN_DESCRIPTION}
  \item{\code{p}}{character COLUMN_DESCRIPTION}
  \item{\code{Uncorrelated.with.PAV.(r2≥0.1)}}{character COLUMN_DESCRIPTION}
  \item{\code{Significant.after.adjusting.for.PAVs}}{character COLUMN_DESCRIPTION}
  \item{\code{Is.a.cis.eQTL.for.same.gene?}}{character COLUMN_DESCRIPTION}
}
}
\usage{
st4
}
\description{
Supplementary information for Sun et al. (2018) Nature
}
\details{
As above.
}
\keyword{datasets}
