% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{st6}
\alias{st6}
\title{Supplementary table 6}
\format{
A data frame with 163 rows and 20 variables:
\describe{
  \item{\code{Locus.ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Sentinel.variant*}}{character COLUMN_DESCRIPTION}
  \item{\code{Chr}}{double COLUMN_DESCRIPTION}
  \item{\code{Pos}}{double COLUMN_DESCRIPTION}
  \item{\code{SOMAmer.ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Target}}{character COLUMN_DESCRIPTION}
  \item{\code{Target.fullname}}{character COLUMN_DESCRIPTION}
  \item{\code{UniProt}}{character COLUMN_DESCRIPTION}
  \item{\code{cis/.trans}}{character COLUMN_DESCRIPTION}
  \item{\code{Mapped.gene}}{character COLUMN_DESCRIPTION}
  \item{\code{Effect.Allele.(EA)}}{character COLUMN_DESCRIPTION}
  \item{\code{Other.Allele.(OA)}}{character COLUMN_DESCRIPTION}
  \item{\code{Previously.reported}}{double COLUMN_DESCRIPTION}
  \item{\code{beta}}{double COLUMN_DESCRIPTION}
  \item{\code{SE}}{double COLUMN_DESCRIPTION}
  \item{\code{p}}{character COLUMN_DESCRIPTION}
  \item{\code{beta}}{double COLUMN_DESCRIPTION}
  \item{\code{SE}}{double COLUMN_DESCRIPTION}
  \item{\code{p}}{character COLUMN_DESCRIPTION}
  \item{\code{Replicates?}}{character COLUMN_DESCRIPTION}
}
}
\usage{
st6
}
\description{
Supplementary information for Sun et al. (2018) Nature
}
\details{
As above.
}
\keyword{datasets}
