% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TwoSampleMR.R
\name{swap}
\alias{swap}
\title{FUNCTION_TITLE}
\usage{
swap(x, y)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\keyword{internal}
