% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{swath_ms}
\alias{swath_ms}
\title{DATASET_TITLE}
\format{
A data frame with 684 rows and 5 variables:
\describe{
  \item{\code{Accession}}{UniProt id}
  \item{\code{accList}}{List of UniProt ids}
  \item{\code{uniprotName}}{Protein}
  \item{\code{ensGene}}{ENSEMBL gene}
  \item{\code{geneName}}{HGNC symbol}
}
}
\usage{
swath_ms
}
\description{
Curated during INTERVAL pilot study.
}
\details{
As above.
}
\keyword{datasets}
