% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turboman.r
\name{turboman}
\alias{turboman}
\title{Manhattan plots for impatient people}
\usage{
turboman(
  input_data_path,
  output_data_rootname,
  custom_peak_annotation_file_path,
  reference_file_path,
  pvalue_sign,
  plot_title
)
}
\arguments{
\item{input_data_path}{Path of the input association data}

\item{output_data_rootname}{Root name of the plot output file}

\item{custom_peak_annotation_file_path}{Path of the custom annotation of variants}

\item{reference_file_path}{Path to the "turboman_hg19_reference_data.rda" reference file}

\item{pvalue_sign}{Significance threshold p-value}

\item{plot_title}{Plot title which will be displayed on top of the plot}
}
\description{
Manhattan plots for impatient people
}
\section{OPTIONAL}{
 Custom annotation file / custom_peak_annotation_file_path
Define the path of the custom annotation of variants...
The input data needs to be a file that has:
\enumerate{
\item Spaces / tabs as field separators
\item One header line with exact column names (order not important)
\item 3 columns: chromosome, position, label (e.g., gene name)
Column names chromosome, position, nearest_gene_name
NB!: If no label is given, variants will be automatically annotated
}
}

\examples{
\dontrun{
require(gap.datasets)
test <- mhtdata[c("chr","pos","p")]
write.table(test,file="test.txt",row.names=FALSE,quote=FALSE)
annotate <- unique(subset(mhtdata[c("chr","start","gene","p")],p<5e-8 & gene!=""))
names(annotate) <- c("chromosome","position","nearest_gene_name","p")
write.table(annotate[,-4],file="annotate.txt",row.names=FALSE,quote=FALSE)
input_data_path <- test.txt
output_data_rootname <- test_man
custom_peak_annotation_file_path <- annotate.txt
reference_file_path <- turboman_hg19_reference_data.rda
pvalue_sign <- 5e-8
plot_title <- "gap.datasets example"
turboman(input_data_path, output_data_rootname, custom_peak_annotation_file_path, reference_file_path, pvalue_sign, plot_title)
}

}
\references{
Berisa et.al, PMID 26395773
}
\author{
Arthur Gilly, Chris Finan, Bram Prins
}
