% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turboqq.r
\name{turboqq}
\alias{turboqq}
\title{An R script to produce QQ plots for impatient people}
\usage{
turboqq(input_data_path, output_data_rootname, plot_title)
}
\arguments{
\item{input_data_path}{Path of the input association data.}

\item{output_data_rootname}{Root name of the plot output file.}

\item{plot_title}{Plot title to be displayed on top of the plot.}
}
\value{
No direct return value. The script generates QQ plots as output.
}
\description{
The method uses the kth order statistic from a sample of n i.i.d. U(0,1) statistics has a Beta(k,n+1-k) distribution.
}
\section{Input association data path / input_data_path}{

Define path of the input association data. The input data needs to be a file that has:
\enumerate{
\item Spaces as field separators
\item One header line
\item Option I (no extreme p-values present): 3 columns, being chromosome, position, pvalue - in this order, column names are not important
Option II (extreme p-values present): 5 columns, being chromosome, position, pvalue, beta, se - in this order, column names are not important
}
}

\examples{
\dontrun{
require(gap.datasets)
test <- mhtdata[c("chr","pos","p")]
write.table(test,file="test.txt",row.names=FALSE,quote=FALSE)
input_data_path <- "test.txt"
output_data_rootname <- "test_qq"
plot_title <- "gap.datasets example"
turboqq(input_data_path, output_data_rootname, plot_title)
}

}
\references{
Quesenberry, C. P., & Hale, D. K. (1980). Journal of Statistical Computation and Simulation, 11(1), 41-53. Coded by
Weale M, Price T. \href{https://sites.google.com/site/mikeweale/software/qqplots/qq_plot_v7.r}{}
}
\author{
Bram Prins
}
