% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniprot2ids}
\alias{uniprot2ids}
\title{UniProt IDs to others}
\usage{
uniprot2ids(uniprotid = "ACC+ID", to, query)
}
\arguments{
\item{uniprotid}{Source IDs.}

\item{to}{To IDs.}

\item{query}{A query.}
}
\value{
A UniProt-ID mapping
}
\description{
A function which converts UniProt IDs to others.
}
\details{
This function is based on the Python3 script from UniProt.
See https://www.uniprot.org/help/api_idmapping
}
\note{
Adapted from script by UniProt
}
\examples{
\dontrun{
uniprotid <- "ACC+ID"
to <- "CHEMBL_ID"
query <- noquote(gap.datasets::inf1[["uniprot"]])
query <- paste(query,collapse=" ")
r <- pQTLtools::uniprot2ids(uniprotid,to,query)
cat(r,file="INF1.merge.chembl")
}
}
\keyword{utilities}
