% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csq.R
\name{csq}
\alias{csq}
\title{Variant consequence}
\usage{
csq(
  query_loci,
  annotated_loci,
  pattern,
  ldops = NULL,
  flanking = 1e+06,
  pop = "EUR",
  verbose = TRUE
)
}
\arguments{
\item{query_loci}{A data.frame of loci whose consequences are to be obtained.}

\item{annotated_loci}{A data.frame of annotated loci.}

\item{ldops}{arguments for ieugwasr::ld_matrix_local()}

\item{flanking}{A flanking distance.}

\item{pop}{The reference population as for ieugwasr::ld_matrix().}

\item{verbose}{A flag to show nonexistent variants.}

\item{consequences}{A set of consequences to check against.}
}
\value{
A data.frame with an indicator with/without those consequences.
}
\description{
This function maps consequences (CSQ) of a given list of variants to those in an annotated
set based on facilities such as variant effect predictor (VEP). In the case of
protein-altering variants (PAVs), many consequences can be involved. The procedure also
takes into account linkage disequilibrium (LD) in flanking regions.
}
