% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\name{get.prop.below.LLOD}
\alias{get.prop.below.LLOD}
\title{Limit of detection analysis}
\usage{
get.prop.below.LLOD(eset, flagged = "OUT")
}
\arguments{
\item{eset}{An ExpressionSet object.}

\item{flagged}{A flag is an indicator for sample exclusion.}
}
\value{
An updated ExpressionSet object.
}
\description{
The function obtains lower limit of detection as in proteomic analysis.
}
\examples{
suppressMessages(library(Biobase))
data(sample.ExpressionSet, package="Biobase")
exampleSet <- sample.ExpressionSet
Biobase::fData(exampleSet)
Biobase::fData(exampleSet)$lod.max <-
    apply(Biobase::exprs(exampleSet),1,quantile,runif(nrow(exampleSet)))
lod <- get.prop.below.LLOD(exampleSet)
x <- dplyr::arrange(fData(lod),desc(pc.belowLOD.new))
knitr::kable(head(lod))
plot(x[,2], main="Random quantile cut off", ylab="<lod\%")
}
\author{
James Peters
}
