% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptideAssociationPlot.R
\name{peptideAssociationPlot}
\alias{peptideAssociationPlot}
\title{peptide association plot}
\usage{
peptideAssociationPlot(protein, cistrans, chrlen = gap::hg19, disp = 85)
}
\arguments{
\item{protein}{protein name.}

\item{cistrans}{peptide signals and classification.}

\item{chrlen}{chromosome length (hg18, hg19, hg38).}

\item{disp}{distance from x-axis.}
}
\value{
None.
}
\description{
peptide association plot
}
\examples{
\dontrun{
  par(mfrow=c(2,1))
  protein <- "PROC"
  suffix <- "_dr"
  input <- paste0("~/Caprion/analysis/METAL",suffix,"/gz/",protein,suffix,".txt.gz")
  annotation <- paste0("~/Caprion/analysis/METAL",suffix,"/vep/",protein,suffix,".txt")
  reference <- file.path(find.package("pQTLtools"),"turboman",
                                      "turboman_hg19_reference_data.rda")
  pvalue_sign <- 5e-8
  plot_title <- protein
  pQTLtools::turboman(input, annotation, reference, pvalue_sign, plot_title)
  cistrans <- read.csv(paste0("~/pQTLtools/tests","/",protein,".cis.vs.trans"))
  load("~/pQTLtools/tests/PROC.rda")
  peptideAssociationPlot(protein,cistrans)
}
}
