% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pqtlMR}
\alias{pqtlMR}
\title{Basic pQTL-MR analysis}
\usage{
pqtlMR(
  ivs,
  outcome,
  mr_plot = FALSE,
  prefix = "pQTL-combined-",
  reverse = FALSE
)
}
\arguments{
\item{ivs}{Instrumental variables from \code{format_data()}.}

\item{outcome}{The counterpart for outcome.}

\item{mr_plot}{to produce plots.}

\item{prefix}{a prefix for output files.}

\item{reverse}{if TRUE, perform reverse MR.}
}
\value{
No value is returned but several files.
}
\description{
This function takes data intrumental variables as produced by \code{format_data()} and
used to perform MR analysis against a list of outcomes from MR-Base.
}
\details{
This function is adapted from \insertCite{zheng20;textual}{pQTLtools} and in
spirit similar to \code{run_TwoSampleMR}.
}
\note{
Adapted from script by Jie Zheng.
}
\examples{
library(TwoSampleMR)
library(pQTLtools)
fi <- file.path(find.package("pQTLtools",lib.loc=.libPaths()),"tests","Ins.csv")
ivs <- format_data(read.csv(fi),type="exposure")
fo <- file.path(find.package("pQTLtools",lib.loc=.libPaths()),"tests","Out.csv")
outcome <- format_data(read.csv(fo),type="outcome")
pqtlMR(ivs, outcome, prefix="IL6R-")
pqtlMR(ivs, outcome, prefix="IL6R_rev-",reverse=TRUE)
# Phenotype,SNP,effect_allele,other_allele,eaf,beta,se,pval
# ABO,rs505922,C,T,0.313,1.298,0.014,1.2e-1828
# LIFR,rs635634,T,C,0.180,-0.300,0.032,6.00E-21
# ids <- c("ieu-a-7","ebi-a-GCST007432")
# f <- file.path(find.package("pQTLtools",lib.loc=.libPaths()),"tests","ms.ins")
# ivs <- format_data(read.table(f, header=TRUE), samplesize_col="N")
# SNP Phenotype effect_allele other_allele eaf beta se pval N
# rs1800693 TNFB T C 0.6033 0.0282  0.0136 0.0389045   11787
# rs2364485 TNFB A C 0.1645 6514963 0.1759 1.62181e-20 11344
# https://raw.githubusercontent.com/MRCIEU/epigraphdb-pqtl/master/scripts/MR-pQTL-script.R
# outcome_data <- TwoSampleMR::extract_outcome_data(snps=with(ivs,SNP),outcomes=ids)

}
\references{
\insertAllCited{}
}
\keyword{utilities}
