% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pQTLtools.R
\docType{package}
\name{pQTLtools}
\alias{pQTLtools}
\alias{pQTLtools-package}
\title{A reference manual}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

It collects data and utilities for pQTL analysis, including 1. Protein GWAS facilities such as Manhattan/QQ/LocusZoom.js plots, novelty/consequence checking; 2. Articles linking functions for cis/trans classification, pQTL-gene plot, 2d/3d-plotly plots, forest plots among others available from 'gap' (\url{https://cran.r-project.org/package=gap}) as well as colocalization, pQTL-Mendelian Randomization via 'TwoSampleMR' (\url{https://mrcieu.github.io/TwoSampleMR/}); 3. Query on genes, regions, and SNPs via 'PhenoScanner' (\url{https://github.com/phenoscanner/}). 4. Mapping of UniProt IDs to other resources; 5. Showcases of 'Bioconductor' (\url{https://github.com/bioconductor}) and 'snakemake' (\url{https://github.com/snakemake}).
}
\details{
Available data and functions are listed in the following table.\tabular{ll}{
   Objects \tab Description \cr
   \strong{GWAS} \tab   \cr
   \code{\link{turboman}} \tab Manhattan plots \cr
   \code{\link{turboqq}} \tab QQ plots \cr
     \tab   \cr
   \strong{pGWAS} \tab   \cr
   \code{\link{csq}} \tab Variant consequence \cr
   \code{\link{novelty_check}} \tab Locus novelty check \cr
   \code{\link{peptideAssociationPlot}} \tab peptide association plot \cr
   \code{\link{peptideMapping}} \tab peptide-to-protein mapping \cr
   \code{\link{protein_altering_variants}} \tab Protein Altering Variants (PAVs) \cr
     \tab   \cr
   \strong{Expression analysis} \tab   \cr
   \code{\link{get.prop.below.LLOD}} \tab Limit of detection analysis \cr
   \code{\link{import_eQTLCatalogue}} \tab Import eQTL Catalogue \cr
   \code{\link{make_ExpressionSet}} \tab A call to ExpressionSet class \cr
   \code{\link{run_coloc}} \tab Colocalisation analysis \cr
     \tab   \cr
   \strong{MR analysis} \tab   \cr
   \code{\link{import_OpenGWAS}} \tab Import OpenGWAS \cr
   \code{\link{pqtlMR}} \tab Bidirectional pQTL-MR analysis \cr
   \code{\link{qtl_lookup}} \tab QTL lookup \cr
   \code{\link{run_TwoSampleMR}} \tab A generic wrapper for TwoSampleMR analysis \cr
     \tab   \cr
   \strong{PhenoScanner Utilities} \tab   \cr
   \code{\link{genequeries}} \tab phenoscanner genequeries in batches \cr
   \code{\link{regionqueries}} \tab phenoscanner regionqueries in batches \cr
   \code{\link{snpqueries}} \tab phenoscanner snpqueries in batches \cr
     \tab   \cr
   \strong{UniProt API} \tab   \cr
   \code{\link{uniprot2ids}} \tab UniProt ID to others \cr
     \tab   \cr
   \strong{Functions in gap} \tab   \cr
   \code{\link[gap:METAL_forestplot]{gap::METAL_forestplot}} \tab Forest plots from metal analysis \cr
   \code{\link[gap:ab]{gap::ab}} \tab Test and Power Calculation for Mediating Effect \cr
   \code{\link[gap:asplot]{gap::asplot}} \tab Regional association plot \cr
   \code{\link[gap:ci2ms]{gap::ci2ms}} \tab Effect size and standard error from confidence interval \cr
   \code{\link[gap:cis.vs.trans.classification]{gap::cis.vs.trans.classification}} \tab a cis/trans classifier \cr
   \code{\link[gap:circos.cis.vs.trans.plot]{gap::circos.cis.vs.trans.plot}} \tab circos plot of cis/trans classification \cr
   \code{\link[gap:circos.mhtplot]{gap::circos.mhtplot}} \tab circos Manhattan plot with gene annotation \cr
   \code{\link[gap:circos.mhtplot2]{gap::circos.mhtplot2}} \tab Another circos Manhattan plot \cr
   \code{\link[gap:cs]{gap::cs}} \tab Credible set \cr
   \code{\link[gap:get_b_se]{gap::get_b_se}} \tab Get b and se from AF, n, and z \cr
   \code{\link[gap:get_pve_se]{gap::get_pve_se}} \tab Get pve and its standard error from n, z \cr
   \code{\link[gap:get_sdy]{gap::get_sdy}} \tab Get sd(y) from AF, n, b, se \cr
   \code{\link[gap:mr]{gap::mr}} \tab Mendelian randomization analysis \cr
   \code{\link[gap:invnormal]{gap::invnormal}} \tab Inverse normal transformation \cr
   \code{\link[gap:log10p]{gap::log10p}} \tab log10(p) for a standard normal deviate \cr
   \code{\link[gap:log10pvalue]{gap::log10pvalue}} \tab log10(p) for a P value including its scientific format \cr
   \code{\link[gap:logp]{gap::logp}} \tab log(p) for a normal deviate \cr
   \code{\link[gap:makeRLEplot]{gap::makeRLEplot}} \tab Create a Relative Log Expression (RLE) Plot \cr
   \code{\link[gap:masize]{gap::masize}} \tab Sample Size Calculation for Mediation Analysis \cr
   \code{\link[gap:mhtplot.trunc]{gap::mhtplot.trunc}} \tab Truncated Manhattan plot \cr
   \code{\link[gap:miamiplot2]{gap::miamiplot2}} \tab Miami plot \cr
   \code{\link[gap:mr_forestplot]{gap::mr_forestplot}} \tab Mendelian Randomization forest plot \cr
   \code{\link[gap:pvalue]{gap::pvalue}} \tab P value for a normal deviate \cr
   \code{\link[gap:qtlClassifier]{gap::qtlClassifier}} \tab A QTL cis/trans classifier \cr
   \code{\link[gap:qtlFinder]{gap::qtlFinder}} \tab Distance-based signal identification \cr
   \code{\link[gap:qtl2dplot]{gap::qtl2dplot}} \tab 2D QTL plot \cr
   \code{\link[gap:qtl2dplotly]{gap::qtl2dplotly}} \tab 2D QTL plotly \cr
   \code{\link[gap:qtl3dplotly]{gap::qtl3dplotly}} \tab 3D QTL plotly \cr
}
}
\section{Usage}{

Vignettes on package usage:
\itemize{
\item An Overview of pQTLtools. \code{vignette("pQTLtools")}.
\item Bioconductor Notes. \code{vignette("bioconductor")}.
\item LocusZoom.js. \code{vignette("LocusZoom.js")}.
\item snakemake showcases. \code{vignette("snakemake")}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jinghuazhao/pQTLtools/}
  \item \url{https://jinghuazhao.github.io/pQTLtools/}
  \item Report bugs at \url{https://github.com/jinghuazhao/pQTLtools/issues}
}

}
\author{
Jing Hua Zhao in collaboration with other colleagues
}
\keyword{internal}
