% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pQTLtools.R
\docType{package}
\name{pQTLtools}
\alias{pQTLtools}
\alias{_PACKAGE}
\alias{pQTLtools-package}
\title{A summary of functions}
\description{
It seeds collection of data and utilities for pQTL analysis. At this
early stage, the repository contains 1. Articles linking functions
for cis/trans classification, 2d Manhattan plots, 2d/3d-plotly plots,
forest plots among others available from
gap, \url{https://github.com/jinghuazhao/R/tree/master/gap};
2. Query on genes, regions, and SNPs via PhenoScanner,
\url{http://www.phenoscanner.medschl.cam.ac.uk/}, adding functionality
to check for replication across platforms; 3. Downstream analysis
such as colocalization, pQTL-Mendelian Randomization via TwoSampleMR,
\url{https://github.com/MRCIEU/TwoSampleMR}, linkage through UniProt IDs
to other resources; 4. Bioconductor notes and a showcase of snakemake
workflow.
}
\details{
Available functions are listed in the following table.\tabular{ll}{
   Objects \tab Description \cr
   \strong{eQTL/GWAS} \tab  \cr
}


\code{get.prop.below.LLOD}  |   Limit of detection analysis
\code{import_eQTLCatalogue} |   Import eQTL Catalogue
\code{import_OpenGWAS}      |   Import OpenGWAS
\code{make_ExpressionSet}   |   A call to ExpressionSet class
\code{novelty_check}        |   Locus novelty check
\code{run_coloc}            |   Colocalisation analysis

\strong{MR analysis}       |

\code{pqtlMR}               |   Bidirectional pQTL-MR analysis
\code{qtl_lookup}           |   QTL lookup
\code{run_TwoSampleMR}      |   A generic wrapper for TwoSampleMR analysis

\strong{PhenoScanner Utilities} |

\code{genequeries}          |   phenoscanner genequeries in batches
\code{regionqueries}        |   phenoscanner regionqueries in batches
\code{snpqueries}           |   phenoscanner snpqueries in batches

\strong{UniProt API}        |

\code{uniprot2ids}          |   UniProt ID to others

\strong{Functions in \code{\link[gap:gap]{gap::gap}}} |

\code{\link[gap:METAL_forestplot]{gap::METAL_forestplot}}    |    Forest plots from metal analysis
\code{\link[gap:ci2ms]{gap::ci2ms}}               |    Effect size and standard error from confidence interval
\code{\link[gap:cis.vs.trans.classification]{gap::cis.vs.trans.classification}} | a cis/trans classifier
\code{\link[gap:circos.cis.vs.trans.plot]{gap::circos.cis.vs.trans.plot}} | circos plot of cis/trans classification
\code{\link[gap:circos.mhtplot]{gap::circos.mhtplot}}      |    circos Manhattan plot with gene annotation
\code{\link[gap:circos.mhtplot2]{gap::circos.mhtplot2}}     |    Another circos Manhattan plot
\code{\link[gap:cs]{gap::cs}}                  |    Credible set
\code{\link[gap:get_b_se]{gap::get_b_se}}            |    Get b and se from AF, n, and z
\code{\link[gap:get_pve_se]{gap::get_pve_se}}          |    Get pve and its standard error from n, z
\code{\link[gap:get_sdy]{gap::get_sdy}}             |    Get sd(y) from AF, n, b, se
\code{\link[gap:mr]{gap::mr}}                  |    Mendelian randomization analysis
\code{\link[gap:invnormal]{gap::invnormal}}           |    Inverse normal transformation
\code{\link[gap:log10p]{gap::log10p}}              |    log10(p) for a standard normal deviate
\code{\link[gap:log10pvalue]{gap::log10pvalue}}         |    log10(p) for a P value including its scientific format
\code{\link[gap:logp]{gap::logp}}                |    log(p) for a normal deviate
\code{\link[gap:mhtplot.trunc]{gap::mhtplot.trunc}}       |    Truncated Manhattan plot
\code{\link[gap:miamiplot2]{gap::miamiplot2}}          |    Miami plot
\code{\link[gap:mr_forestplot]{gap::mr_forestplot}}       |    Mendelian Randomization forest plot
\code{\link[gap:qtlClassifier]{gap::qtlClassifier}}       |    A QTL cis/trans classifier
\code{\link[gap:qtl2dplot]{gap::qtl2dplot}}           |    2D QTL plot
\code{\link[gap:qtl2dplotly]{gap::qtl2dplotly}}         |    2D QTL plotly
\code{\link[gap:qtl3dplotly]{gap::qtl3dplotly}}         |    3D QTL plotly
}
\section{Usage}{

Vignettes on package usage:
\itemize{
\item An Overview of pQTLtools. \code{vignette("pQTLtools")}.
\item Bioconductor Notes. \code{vignette("bioconductor")}.
\item ExpressionSet usage. \code{vignette("es")}.
\item snakemake showcases. \code{vignette("snakemake")}.
\item SCALLOP-INF scripts. \code{vignette("SCALLOP-INF")}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jinghuazhao/pQTLtools}
  \item Report bugs at \url{https://github.com/jinghuazhao/pQTLtools/issues}
}

}
\author{
Jing Hua Zhao in collaboration with other colleagues
}
\keyword{internal}
