% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pQTLtools.R
\docType{package}
\name{pQTLtools}
\alias{pQTLtools}
\alias{_PACKAGE}
\alias{pQTLtools-package}
\title{A summary of functions}
\description{
It seeds collection of data and utilities for pQTL analysis. At this
early stage, the repository contains 1. Articles linking functions
for cis/trans classification, 2d Manhattan plots, 2d/3d-plotly plots,
forest plots among others available from
gap, \url{https://github.com/jinghuazhao/R/tree/master/gap};
2. Query on genes, regions, and SNPs via PhenoScanner,
\url{http://www.phenoscanner.medschl.cam.ac.uk/}, adding functionality
to check for replication across platforms; 3. Downstream analysis
such as colocalization, pQTL-Mendelian Randomization via TwoSampleMR,
\url{https://github.com/MRCIEU/TwoSampleMR}, linkage through UniProt IDs
to other resources; 4. Bioconductor notes and a showcase of snakemake
workflow.
}
\details{
Available functions are listed in the following table.

\tabular{ll}{
Objects              &   Description\cr
\cr
\strong{eQTL/GWAS}\cr
\cr
\code{get.prop.below.LLOD}  &   Limit of detection analysis\cr
\code{import_eQTLCatalogue} &   Import eQTL Catalogue\cr
\code{import_OpenGWAS}      &   Import OpenGWAS\cr
\code{make_ExpressionSet}   &   A call to ExpressionSet class\cr
\code{novelty_check}        &   Locus novelty check\cr
\code{run_coloc}            &   Colocalisation analysis\cr
\cr
\strong{MR analysis}\cr
\cr
\code{pqtlMR}               &   Bidirectional pQTL-MR analysis\cr
\code{qtl_lookup}           &   QTL lookup\cr
\code{run_TwoSampleMR}      &   A generic wrapper for TwoSampleMR analysis\cr
\cr
\strong{PhenoScanner Utilities}\cr
\cr
\code{genequeries}          &   phenoscanner genequeries in batches\cr
\code{regionqueries}        &   phenoscanner regionqueries in batches\cr
\code{snpqueries}           &   phenoscanner snpqueries in batches\cr
\cr
\strong{UniProt API}   \cr
\cr
\code{uniprot2ids}          &   UniProt ID to others\cr
\cr
\strong{Functions in gap}\cr
\cr
\code{\link[gap:METAL_forestplot]{gap::METAL_forestplot}}    &    Forest plots from metal analysis\cr
\code{\link[gap:ci2ms]{gap::ci2ms}}               &    Effect size and standard error from confidence interval\cr
\code{\link[gap:cis.vs.trans.classification]{gap::cis.vs.trans.classification}} & a cis/trans classifier\cr
\code{\link[gap:circos.cis.vs.trans.plot]{gap::circos.cis.vs.trans.plot}} & circos plot of cis/trans classification\cr
\code{\link[gap:circos.mhtplot]{gap::circos.mhtplot}}      &    circos Manhattan plot with gene annotation\cr
\code{\link[gap:circos.mhtplot2]{gap::circos.mhtplot2}}     &    Another circos Manhattan plot\cr
\code{\link[gap:cs]{gap::cs}}                  &    Credible set\cr
\code{\link[gap:get_b_se]{gap::get_b_se}}            &    Get b and se from AF, n, and z\cr
\code{\link[gap:get_pve_se]{gap::get_pve_se}}          &    Get pve and its standard error from n, z\cr
\code{\link[gap:get_sdy]{gap::get_sdy}}             &    Get sd(y) from AF, n, b, se\cr
\code{\link[gap:mr]{gap::mr}}                  &    Mendelian randomization analysis\cr
\code{\link[gap:invnormal]{gap::invnormal}}           &    Inverse normal transformation\cr
\code{\link[gap:log10p]{gap::log10p}}              &    log10(p) for a standard normal deviate\cr
\code{\link[gap:log10pvalue]{gap::log10pvalue}}         &    log10(p) for a P value including its scientific format\cr
\code{\link[gap:logp]{gap::logp}}                &    log(p) for a normal deviate\cr
\code{\link[gap:mhtplot.trunc]{gap::mhtplot.trunc}}       &    Truncated Manhattan plot\cr
\code{\link[gap:miamiplot2]{gap::miamiplot2}}          &    Miami plot\cr
\code{\link[gap:mr_forestplot]{gap::mr_forestplot}}       &    Mendelian Randomization forest plot\cr
\code{\link[gap:qtlClassifier]{gap::qtlClassifier}}       &    A QTL cis/trans classifier\cr
\code{\link[gap:qtl2dplot]{gap::qtl2dplot}}           &    2D QTL plot\cr
\code{\link[gap:qtl2dplotly]{gap::qtl2dplotly}}         &    2D QTL plotly\cr
\code{\link[gap:qtl3dplotly]{gap::qtl3dplotly}}         &    3D QTL plotly
}
}
\section{Usage}{

Vignettes on package usage:
\itemize{
\item An Overview of pQTLtools. \code{vignette("pQTLtools")}.
\item Bioconductor Notes. \code{vignette("bioconductor")}.
\item ExpressionSet usage. \code{vignette("es")}.
\item snakemake showcases. \code{vignette("snakemake")}.
\item SCALLOP-INF scripts. \code{vignette("SCALLOP-INF")}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jinghuazhao/pQTLtools}
  \item Report bugs at \url{https://github.com/jinghuazhao/pQTLtools/issues}
}

}
\author{
Jing Hua Zhao in collaboration with other colleagues
}
\keyword{internal}
