% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pQTLtools.R
\docType{package}
\name{pQTLtools}
\alias{pQTLtools}
\alias{_PACKAGE}
\alias{pQTLtools-package}
\title{A summary of functions}
\description{
It seeds collection of data and utilities for pQTL analysis. At this
early stage, the repository contains 1. Articles linking functions
for cis/trans classification, 2d Manhattan plots, 2d/3d-plotly plots,
forest plots among others available from
gap, \url{https://github.com/jinghuazhao/R/tree/master/gap};
2. Query on genes, regions, and SNPs via PhenoScanner,
\url{http://www.phenoscanner.medschl.cam.ac.uk/}, adding functionality
to check for replication across platforms; 3. Downstream analysis
such as colocalization, pQTL-Mendelian Randomization via TwoSampleMR,
\url{https://github.com/MRCIEU/TwoSampleMR}, linkage through UniProt IDs
to other resources; 4. Bioconductor notes and a showcase of snakemake
workflow.
}
\details{
Available functions are listed in the following table.\tabular{ll}{
   Objects \tab Description \cr
   \strong{eQTL/GWAS} \tab   \cr
   \code{get.prop.below.LLOD} \tab Limit of detection analysis \cr
   \code{import_eQTLCatalogue} \tab Import eQTL Catalogue \cr
   \code{import_OpenGWAS} \tab Import OpenGWAS \cr
   \code{make_ExpressionSet} \tab A call to ExpressionSet class \cr
   \code{novelty_check} \tab Locus novelty check \cr
   \code{run_coloc} \tab Colocalisation analysis \cr
     \tab   \cr
   \strong{MR analysis} \tab   \cr
   \code{pqtlMR} \tab Bidirectional pQTL-MR analysis \cr
   \code{qtl_lookup} \tab QTL lookup \cr
   \code{run_TwoSampleMR} \tab A generic wrapper for TwoSampleMR analysis \cr
     \tab   \cr
   \strong{PhenoScanner Utilities} \tab   \cr
   \code{genequeries} \tab phenoscanner genequeries in batches \cr
   \code{regionqueries} \tab phenoscanner regionqueries in batches \cr
   \code{snpqueries} \tab phenoscanner snpqueries in batches \cr
     \tab   \cr
   \strong{UniProt API} \tab   \cr
   \code{uniprot2ids} \tab UniProt ID to others \cr
     \tab   \cr
   \strong{Functions in \code{\link[gap:gap]{gap::gap}}} \tab   \cr
   \code{\link[gap:METAL_forestplot]{gap::METAL_forestplot}} \tab Forest plots from metal analysis \cr
   \code{\link[gap:ci2ms]{gap::ci2ms}} \tab Effect size and standard error from confidence interval \cr
   \code{\link[gap:cis.vs.trans.classification]{gap::cis.vs.trans.classification}} \tab a cis/trans classifier \cr
   \code{\link[gap:circos.cis.vs.trans.plot]{gap::circos.cis.vs.trans.plot}} \tab circos plot of cis/trans classification \cr
   \code{\link[gap:circos.mhtplot]{gap::circos.mhtplot}} \tab circos Manhattan plot with gene annotation \cr
   \code{\link[gap:circos.mhtplot2]{gap::circos.mhtplot2}} \tab Another circos Manhattan plot \cr
   \code{\link[gap:cs]{gap::cs}} \tab Credible set \cr
   \code{\link[gap:get_b_se]{gap::get_b_se}} \tab Get b and se from AF, n, and z \cr
   \code{\link[gap:get_pve_se]{gap::get_pve_se}} \tab Get pve and its standard error from n, z \cr
   \code{\link[gap:get_sdy]{gap::get_sdy}} \tab Get sd(y) from AF, n, b, se \cr
   \code{\link[gap:mr]{gap::mr}} \tab Mendelian randomization analysis \cr
   \code{\link[gap:invnormal]{gap::invnormal}} \tab Inverse normal transformation \cr
   \code{\link[gap:log10p]{gap::log10p}} \tab log10(p) for a standard normal deviate \cr
   \code{\link[gap:log10pvalue]{gap::log10pvalue}} \tab log10(p) for a P value including its scientific format \cr
   \code{\link[gap:logp]{gap::logp}} \tab log(p) for a normal deviate \cr
   \code{\link[gap:mhtplot.trunc]{gap::mhtplot.trunc}} \tab Truncated Manhattan plot \cr
   \code{\link[gap:miamiplot2]{gap::miamiplot2}} \tab Miami plot \cr
   \code{\link[gap:mr_forestplot]{gap::mr_forestplot}} \tab Mendelian Randomization forest plot \cr
   \code{\link[gap:qtlClassifier]{gap::qtlClassifier}} \tab A QTL cis/trans classifier \cr
   \code{\link[gap:qtl2dplot]{gap::qtl2dplot}} \tab 2D QTL plot \cr
   \code{\link[gap:qtl2dplotly]{gap::qtl2dplotly}} \tab 2D QTL plotly \cr
   \code{\link[gap:qtl3dplotly]{gap::qtl3dplotly}} \tab 3D QTL plotly \cr
}
}
\section{Usage}{

Vignettes on package usage:
\itemize{
\item An Overview of pQTLtools. \code{vignette("pQTLtools")}.
\item Bioconductor Notes. \code{vignette("bioconductor")}.
\item ExpressionSet usage. \code{vignette("es")}.
\item snakemake showcases. \code{vignette("snakemake")}.
\item SCALLOP-INF scripts. \code{vignette("SCALLOP-INF")}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jinghuazhao/pQTLtools}
  \item Report bugs at \url{https://github.com/jinghuazhao/pQTLtools/issues}
}

}
\author{
Jing Hua Zhao in collaboration with other colleagues
}
\keyword{internal}
