% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pav.r
\docType{data}
\name{protein_altering_variants}
\alias{protein_altering_variants}
\title{Protein-altering variants}
\format{
A vector of consequences.
}
\usage{
protein_altering_variants
}
\description{
Variants which directly affect the coding sequence or the protein produced.
}
\examples{
protein_altering_variants
}
\keyword{datasets}
