% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qtl_lookup}
\alias{qtl_lookup}
\title{QTL lookup}
\usage{
qtl_lookup(
  d,
  dat,
  panel = "1000Genomes",
  pthreshold = 0.001,
  pop = "EUR",
  plink_bin = NULL,
  r = NULL,
  r2 = NULL,
  xlsx = NULL
)
}
\arguments{
\item{d}{directory where `dat` (below) is held.}

\item{dat}{MR results (`protein`, `id`, `pqtl`,`p`, `qtl`, `p_qtl`) whose `proxy`, `p_proxy` and `rsq` variables will be updated.}

\item{panel}{reference panel.}

\item{pthreshold}{cutoff of QTL association from trait GWASs.}

\item{pop}{pupulation.}

\item{plink_bin}{PLINK executable file whose binary files are indicated in `bfile` variable of `dat`.}

\item{r}{when specified, the LD(r) is output.}

\item{r2}{when specified, the LD(r^2) is output.}

\item{xlsx}{a non-null specification indicates name of an output Excel workbook.}
}
\value{
A data.frame containing the looked up loci.
}
\description{
This function takes MR results (involving pQTL/QTL) to look up QTLs in trait GWASs given a P value cutoff. 
The rationale is that a pQTL may not necessarily be in strong LD with QTL but some other independent signal
in the same region.
}
\examples{
\dontrun{
options(width=200)
INF <- Sys.getenv("INF")
suppressMessages(library(dplyr))
d <- file.path(INF,"mr","gsmr","trait")
gsmr_efo <- read.delim(file.path(INF,"mr","gsmr","gsmr-efo.txt")) \%>\%
            filter(fdr<=0.05) \%>\%
            mutate(prot=gsub("-",".",protein)) \%>\%
            mutate(file_gwas=paste(prot,id,"rsid.txt",sep="-"),
                   bfile=file.path(INF,"INTERVAL","per_chr",
                                   paste0("interval.imputed.olink.chr_",chr)),
                   proxy=NA,p_proxy=NA,rsq=NA)
proxies <- qtl_lookup(gsmr_efo,plink_bin="/rds/user/jhz22/hpc-work/bin/plink",
                      xlsx=file.path(INF,"mr","gsmr","r2_INTERVAL.xlsx")) \%>\%
           select(protein,id,Disease,pqtl,p,qtl,p_qtl,proxy,p_proxy,rsq)
write.table(proxies,file=file.path(INF,"mr","gsmr","r2_INTERVAL.tsv"),
            row.names=FALSE,quote=FALSE,sep="\t")
}
}
