% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{st4}
\alias{st4}
\title{Supplementary table 4}
\format{
A data frame with 1980 rows and 31 variables:
\describe{
  \item{\code{Locus.ID}}{Locus id}
  \item{\code{SOMAmer.ID}}{SOMAmer id}
  \item{\code{Target}}{Protein}
  \item{\code{Target.fullname}}{Protein full name}
  \item{\code{UniProt}}{UniProt id}
  \item{\code{Sentinel.variant*}}{Sentinel variant}
  \item{\code{Chr}}{Chromosome}
  \item{\code{Pos}}{Position}
  \item{\code{Region.start}}{Region start}
  \item{\code{Region.end}}{Region end}
  \item{\code{Effect.Allele.(EA)}}{Effect allele}
  \item{\code{Other.Allele.(OA)}}{Other allele}
  \item{\code{EAF}}{Effect allele frequency}
  \item{\code{INFO}}{Information score}
  \item{\code{cis/.trans}}{"cis"/"trans"}
  \item{\code{Mapped.gene}}{Mapped gene}
  \item{\code{No..conditionally.significant.variants}}{Number of conditionally significant variants}
  \item{\code{Previously.reported}}{Previously reported}
  \item{\code{Replicates?}}{Yes/No}
  \item{\code{beta}}{b}
  \item{\code{SE}}{SE}
  \item{\code{p}}{p value}
  \item{\code{beta}}{b}
  \item{\code{SE}}{SE}
  \item{\code{p}}{p value}
  \item{\code{beta}}{b}
  \item{\code{SE}}{SE}
  \item{\code{p}}{p value}
  \item{\code{Uncorrelated.with.PAV.(r2≥0.1)}}{Uncorrelated with PAV}
  \item{\code{Significant.after.adjusting.for.PAVs}}{Significant after adjustment for PAVs}
  \item{\code{Is.a.cis.eQTL.for.same.gene?}}{Is a cis eQTL for the same gene?}
}
}
\usage{
st4
}
\description{
Supplementary information for Sun et al. (2018) Nature
}
\details{
As above.
}
\keyword{datasets}
\keyword{internal}
