% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{st6}
\alias{st6}
\title{Supplementary table 6}
\format{
A data frame with 163 rows and 20 variables:
\describe{
  \item{\code{Locus.ID}}{Locus id}
  \item{\code{Sentinel.variant*}}{Sentinel variant}
  \item{\code{Chr}}{Chromosome [1-12,14-20,22]}
  \item{\code{Pos}}{Position}
  \item{\code{SOMAmer.ID}}{SOMAmer id}
  \item{\code{Target}}{Protein}
  \item{\code{Target.fullname}}{Protein full name}
  \item{\code{UniProt}}{UniProt id}
  \item{\code{cis/.trans}}{"cis"/"trans"}
  \item{\code{Mapped.gene}}{Mapped gene}
  \item{\code{Effect.Allele.(EA)}}{Effect allele}
  \item{\code{Other.Allele.(OA)}}{Other allele}
  \item{\code{Previously.reported}}{Previously reported (0/1)}
  \item{\code{beta}}{b}
  \item{\code{SE}}{SE}
  \item{\code{p}}{p value}
  \item{\code{beta}}{b}
  \item{\code{SE}}{SE}
  \item{\code{p}}{p value}
  \item{\code{Replicates?}}{Yes/No}
}
}
\usage{
st6
}
\description{
Supplementary information for Sun et al. (2018) Nature
}
\details{
As above.
}
\keyword{datasets}
