% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turboman.r
\name{turboman}
\alias{turboman}
\title{Manhattan plots}
\usage{
turboman(
  input_data_path,
  custom_peak_annotation_file_path,
  reference_file_path,
  pvalue_sign,
  plot_title
)
}
\arguments{
\item{input_data_path}{Path of the input association data}

\item{custom_peak_annotation_file_path}{Path of the custom annotation of variants}

\item{reference_file_path}{Path to the "turboman_hg19_reference_data.rda" / "turboman_hg19_reference_data.rda" reference file}

\item{pvalue_sign}{Significance threshold p-value}

\item{plot_title}{Plot title which will be displayed on top of the plot}
}
\description{
Manhattan plots
}
\details{
\strong{Input association data path / input_data_path}

Define the path of the input association data...
The input data needs to be a file that has:
\enumerate{
\item Spaces as field separators.
\item One header line.
\item Option I. (no extreme p-values present): 3 columns, being chromosome, position, pvalue in order, column names are not important.
Option II. (extreme p-values present): 5 columns, being chromosome, position, pvalue, beta, se in order, column names are not important.
}

\strong{OPTIONAL: Custom annotation file / custom_peak_annotation_file_path}

Define the path of the custom annotation of variants...
The input data needs to be a file that has:
\enumerate{
\item Spaces / tabs as field separators.
\item One header line with exact column names (order not important).
\item Columns: chromosome, position, label (e.g., gene name) / nearest_gene_name, cis/trans flag (optional).
NB!: If no label is given, variants will be automatically annotated
}

\strong{Reference file path / reference_file_path}

Define the path to the "turboman_hg19_reference_data.rda" / "turboman_hg38_reference_data.rda" reference file that
contains the LD block breaks as in \insertCite{berisa16;textual}{pQTLtools}
and gene coordinates used to construct and annotate the Manhattan plot. Both
are available from the \code{turboman} directory of the installed package, e.g.,
\code{file.path(find.package("pQTLtools"),"turboman","turboman_hg38_reference_data.rda")}.

\strong{Significance threshold p-value / pvalue_sign}

Define the significance threshold...
This will be used to...
\enumerate{
\item Highlight signal peaks that come above this significance threshold.
\item Annotate the nearest gene to the top signal in the peak.
\item Draw a horizontal reference line equal to this threshold.
}
}
\examples{
\dontrun{
require(gap.datasets)
test <- mhtdata[c("chr","pos","p")]
write.table(test,file="test.txt",row.names=FALSE,quote=FALSE)
annotate <- unique(subset(mhtdata[c("chr","start","gene","p")],p<5e-8 & gene!=""))
names(annotate) <- c("chromosome","position","nearest_gene_name","p")
write.table(annotate[,-4],file="annotate.txt",row.names=FALSE,quote=FALSE)
input_data_path <- "test.txt"
output_data_rootname <- "test_man"
custom_peak_annotation_file_path <- "annotate.txt"
reference_file_path <- "turboman_hg19_reference_data.rda"
pvalue_sign <- 5e-8
plot_title <- "gap.datasets example"
turboman(input_data_path, custom_peak_annotation_file_path, reference_file_path, pvalue_sign, plot_title)

input_data_path <- "FCN3.txt.gz"
custom_peak_annotation_file_path <- "FCN3.txt"
reference_file_path <- "turboman_hg19_reference_data.rda"
pvalue_sign <- 5e-8
plot_title <- "FCN3"
png("FCN3.png",width=3600, height=3600, pointsize = 12, res=450)
turboman(input_data_path, custom_peak_annotation_file_path, reference_file_path, pvalue_sign, plot_title)
dev.off()
}

}
\references{
\insertAllCited{}
}
\author{
Arthur Gilly, Chris Finan, Bram Prins, see \href{https://github.com/bpprins/turboman}{https://github.com/bpprins/turboman}.
}
