% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turboqq.r
\name{turboqq}
\alias{turboqq}
\title{QQ plots}
\usage{
turboqq(input_data_path, plot_title, plot_resolution = 1800)
}
\arguments{
\item{input_data_path}{Path of the input association data.}

\item{plot_title}{Plot title to be displayed on top of the plot.}

\item{plot_resolution}{a fixed number of points (pixels) to be plotted vertically and horizontally.}
}
\value{
No direct return value. The script generates QQ plots as output.
}
\description{
QQ plots
}
\details{
The method uses the kth order statistic from a sample of n i.i.d. U(0,1) statistics has a Beta(k,n+1-k) distribution as in
\insertCite{quesenberry80;textual}{pQTLtools} and Coded by  Weale M, Price T.
\href{https://sites.google.com/site/mikeweale/software}{https://sites.google.com/site/mikeweale/software}

\strong{Input association data file / input_data_path:}

Define path of the input association data. The input data needs to be a file that has:
\enumerate{
\item Spaces as field separators.
\item One header line.
\item Option I. (no extreme p-values present): 3 columns, being chromosome, position, pvalue in order, column names are not important.
Option II. (extreme p-values present): 5 columns, being chromosome, position, pvalue, beta, se in order, column names are not important.
}

\strong{Plot title / plot_title}

Define plot title which will be displayed on top of the plot.

\strong{Rsolution of plot / plot_resolution}

Define a fixed number of points (pixels) to be plotted vertically and horizontally.
}
\examples{
\dontrun{
require(gap.datasets)
test <- mhtdata[c("chr","pos","p")]
write.table(test,file="test.txt",row.names=FALSE,quote=FALSE)
input_data_path <- "test.txt"
plot_title <- "gap.datasets example"
png("test_qq.png", height = 1800, width = 1800, pointsize = 12, res = 450)
par(mar = c(4, 4, 3, 1))
turboqq(input_data_path, plot_title)
dev.off()
}

}
\references{
\insertAllCited{}
}
\author{
Bram Prins, \href{https://github.com/bpprins/turboqq}{https://github.com/bpprins/turboqq}.
}
